+++ nanonet_ml/data/__init__.py
@@ -2,12 +2,14 @@
 
 from nanonet_ml.data.dataloader import DataLoader
 from nanonet_ml.data.dataset import Dataset, TensorDataset
+from nanonet_ml.data.log_dataset import LogDataset
 from nanonet_ml.data.mnist import download_mnist, load_mnist
 
 __all__ = [
     "Dataset",
     "TensorDataset",
     "DataLoader",
+    "LogDataset",
     "load_mnist",
     "download_mnist",
 ]